continued work on thirdparty caveats. Think i grock it now.
the authmacaroon in the verify call is only used for request binding and recursive calls to discharge macaroons, so we dont need the copy stuff. It just confused me when I did the initial work for first party caveats. But the initial signature in verify is computed on the current macaroon id, NOT the auth macaroon id as I had before. That made all the "delegated" go away, and makes much more sense now.

I kept in the fluent assertion pattern cause its nice for adding many caveats in a row. at least for now.